requirements.txt
flask, 
openai-whisper, 
deep-translator,
 gunicorn.